+++ src/GOL.Console/GOL.Console.csproj
@@ -7,9 +7,7 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <Compile Remove="Gameoflife.cs" />
-  </ItemGroup>
+
 
   <ItemGroup>
     <ProjectReference Include="..\GOL.Business\GOL.Business.csproj" />
